+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Feedback is now an inline form rather than a mailto link.
+
 ### Fixed
 - Issue with fullscreen functionality that breaks mobile clients.
 - Unsplash settings and fetched image sometimes getting out of sync.